CHE-4288: Frequent 'Unable to initialize IDE' error on AWS
Move the logic for injecting the Orion editor from the
OrionEditorExtension class to the EditorInitializerPromiseHolder. Thus
avoiding the chance for having the initializePromise field with null
value.
